+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ##### Bug Fixes
 
 - Fix false positive when protocols requirements are satisfied in another file from the one that declares the conformance.
+- Fix redundant public accessibility analysis of enum associated value types.
 
 ## 2.8.3 (2021-11-29)